+++ src/dfx-core/src/config/model/local_server_descriptor.rs
@@ -1,9 +1,9 @@
 use crate::config::model::bitcoin_adapter;
 use crate::config::model::canister_http_adapter::HttpAdapterLogLevel;
 use crate::config::model::dfinity::{
-    ConfigDefaultsBitcoin, ConfigDefaultsCanisterHttp, ConfigDefaultsProxy, ConfigDefaultsReplica,
-    DEFAULT_PROJECT_LOCAL_BIND, DEFAULT_SHARED_LOCAL_BIND, ReplicaLogLevel, ReplicaSubnetType,
-    to_socket_addr,
+    ConfigDefaultsBitcoin, ConfigDefaultsCanisterHttp, ConfigDefaultsDogecoin, ConfigDefaultsProxy,
+    ConfigDefaultsReplica, DEFAULT_PROJECT_LOCAL_BIND, DEFAULT_SHARED_LOCAL_BIND, ReplicaLogLevel,
+    ReplicaSubnetType, to_socket_addr,
 };
 use crate::config::model::replica_config::CachedConfig;
 use crate::error::network_config::{
@@ -16,6 +16,7 @@ use serde::{Deserialize, Serialize};
 use slog::{Logger, debug, info};
 use std::net::SocketAddr;
 use std::path::{Path, PathBuf};
+use std::str::FromStr;
 use time::OffsetDateTime;
 
 use super::replica_config::CachedReplicaConfig;
@@ -45,6 +46,7 @@ pub struct LocalServerDescriptor {
     pub bind_address: SocketAddr,
 
     pub bitcoin: ConfigDefaultsBitcoin,
+    pub dogecoin: ConfigDefaultsDogecoin,
     pub canister_http: ConfigDefaultsCanisterHttp,
     pub proxy: ConfigDefaultsProxy,
     pub replica: ConfigDefaultsReplica,
@@ -67,6 +69,7 @@ impl LocalServerDescriptor {
         data_directory: PathBuf,
         bind: String,
         bitcoin: ConfigDefaultsBitcoin,
+        dogecoin: ConfigDefaultsDogecoin,
         canister_http: ConfigDefaultsCanisterHttp,
         proxy: ConfigDefaultsProxy,
         replica: ConfigDefaultsReplica,
@@ -80,6 +83,7 @@ impl LocalServerDescriptor {
             settings_digest,
             bind_address,
             bitcoin,
+            dogecoin,
             canister_http,
             proxy,
             replica,
@@ -184,9 +188,21 @@ impl LocalServerDescriptor {
     }
 
     pub fn with_bitcoin_enabled(self) -> LocalServerDescriptor {
-        let bitcoin = ConfigDefaultsBitcoin {
-            enabled: true,
-            ..self.bitcoin
+        // if we're enabling bitcoin and there are no bitcoin nodes specified,
+        // we will add a default node at 127.0.0.1:18444
+        let bitcoin = if self.bitcoin.nodes.is_none()
+            || self.bitcoin.nodes.clone().is_some_and(|n| n.is_empty())
+        {
+            ConfigDefaultsBitcoin {
+                enabled: true,
+                nodes: Some(vec![SocketAddr::from_str("127.0.0.1:18444").unwrap()]),
+                ..self.bitcoin
+            }
+        } else {
+            ConfigDefaultsBitcoin {
+                enabled: true,
+                ..self.bitcoin
+            }
         };
         Self { bitcoin, ..self }
     }
@@ -197,6 +213,33 @@ impl LocalServerDescriptor {
             ..self.bitcoin
         };
         Self { bitcoin, ..self }
+    }
+
+    pub fn with_dogecoin_enabled(self) -> LocalServerDescriptor {
+        // if we're enabling dogecoin and there are no dogecoin nodes specified,
+        // we will add a default node at 127.0.0.1:18444
+        let dogecoin = if self.dogecoin.nodes.is_none()
+            || self.dogecoin.nodes.clone().is_some_and(|n| n.is_empty())
+        {
+            ConfigDefaultsDogecoin {
+                enabled: true,
+                nodes: Some(vec![SocketAddr::from_str("127.0.0.1:18444").unwrap()]),
+            }
+        } else {
+            ConfigDefaultsDogecoin {
+                enabled: true,
+                ..self.dogecoin
+            }
+        };
+        Self { dogecoin, ..self }
+    }
+
+    pub fn with_dogecoin_nodes(self, nodes: Vec<SocketAddr>) -> LocalServerDescriptor {
+        let dogecoin = ConfigDefaultsDogecoin {
+            nodes: Some(nodes),
+            ..self.dogecoin
+        };
+        Self { dogecoin, ..self }
     }
 
     pub fn with_proxy_domains(self, domains: Vec<String>) -> LocalServerDescriptor {