+++ src/dfx/src/actors/pocketic.rs
@@ -55,6 +55,8 @@ pub struct Config {
     pub replica_config: ReplicaConfig,
     pub bitcoind_addr: Option<Vec<SocketAddr>>,
     pub enable_bitcoin: bool,
+    pub dogecoind_addr: Option<Vec<SocketAddr>>,
+    pub enable_dogecoin: bool,
     pub port: Option<u16>,
     pub port_file: PathBuf,
     pub pid_file: PathBuf,
@@ -279,6 +281,8 @@ fn pocketic_start_thread(
                 &config.effective_config_path,
                 &config.bitcoind_addr,
                 config.enable_bitcoin,
+                &config.dogecoind_addr,
+                config.enable_dogecoin,
                 &config.replica_config,
                 config.pocketic_proxy_config.domains.clone(),
                 config.pocketic_proxy_config.bind,
@@ -342,6 +346,8 @@ async fn initialize_pocketic(
     effective_config_path: &Path,
     bitcoind_addr: &Option<Vec<SocketAddr>>,
     enable_bitcoin: bool,
+    dogecoind_addr: &Option<Vec<SocketAddr>>,
+    enable_dogecoin: bool,
     replica_config: &ReplicaConfig,
     domains: Option<Vec<String>>,
     addr: SocketAddr,
@@ -389,10 +395,12 @@ async fn initialize_pocketic(
             nns_ui: Some(IcpFeaturesConfig::default()),
             bitcoin: icp_features.bitcoin,
             canister_migration: None,
+            dogecoin: icp_features.dogecoin,
         }
     } else {
         icp_features
     };
+
     let icp_features = if bitcoind_addr.is_some() || enable_bitcoin {
         IcpFeatures {
             bitcoin: Some(IcpFeaturesConfig::default()),
@@ -402,29 +410,41 @@ async fn initialize_pocketic(
         icp_features
     };
 
+    let icp_features = if dogecoind_addr.is_some() || enable_dogecoin {
+        IcpFeatures {
+            dogecoin: Some(IcpFeaturesConfig::default()),
+            ..icp_features
+        }
+    } else {
+        icp_features
+    };
+
+    let instance_config = InstanceConfig {
+        subnet_config_set,
+        state_dir: Some(replica_config.state_manager.state_root.clone()),
+        icp_config: Some(IcpConfig {
+            beta_features: Some(IcpConfigFlag::Enabled),
+            ..Default::default()
+        }),
+        log_level: Some(replica_config.log_level.to_pocketic_string()),
+        bitcoind_addr: bitcoind_addr.clone(),
+        dogecoind_addr: dogecoind_addr.clone(),
+        icp_features: Some(icp_features),
+        http_gateway_config: Some(InstanceHttpGatewayConfig {
+            ip_addr: Some(addr.ip().to_string()),
+            port: Some(addr.port()),
+            domains,
+            https_config: None,
+        }),
+        initial_time: Some(InitialTime::AutoProgress(AutoProgressConfig {
+            artificial_delay_ms: Some(replica_config.artificial_delay as u64),
+        })),
+        ..Default::default()
+    };
+
     let resp = init_client
         .post(format!("http://localhost:{port}/instances"))
-        .json(&InstanceConfig {
-            subnet_config_set,
-            state_dir: Some(replica_config.state_manager.state_root.clone()),
-            icp_config: Some(IcpConfig {
-                beta_features: Some(IcpConfigFlag::Enabled),
-                ..Default::default()
-            }),
-            log_level: Some(replica_config.log_level.to_pocketic_string()),
-            bitcoind_addr: bitcoind_addr.clone(),
-            icp_features: Some(icp_features),
-            http_gateway_config: Some(InstanceHttpGatewayConfig {
-                ip_addr: Some(addr.ip().to_string()),
-                port: Some(addr.port()),
-                domains,
-                https_config: None,
-            }),
-            initial_time: Some(InitialTime::AutoProgress(AutoProgressConfig {
-                artificial_delay_ms: Some(replica_config.artificial_delay as u64),
-            })),
-            ..Default::default()
-        })
+        .json(&instance_config)
         .send()
         .await?
         .error_for_status()?
@@ -464,6 +484,8 @@ async fn initialize_pocketic(
 fn initialize_pocketic(
     _: u16,
     _: &Path,
+    _: &Option<Vec<SocketAddr>>,
+    _: bool,
     _: &Option<Vec<SocketAddr>>,
     _: bool,
     _: &ReplicaConfig,