+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # UNRELEASED
 
+# 0.30.1
+
+### feat: support dogecoin for the local dev environment
+
+You can now launch a network with `dfx start --enable-dogeoin` to run the dogecoin
+integration locally.
+
 ### feat: improved the canister snapshot download/upload feature
 
 Improved the canister snapshot download/upload feature by
@@ -15,15 +22,20 @@ Improved the canister snapshot download/upload feature by
 The custom logic was prone to becoming outdated, such as not adapting to changing cycles fees.
 By using `pocket-ic`, which gets updated frequently, the BTC integration is significanly less likely to break.
 
+### fix: `dfx start --enable-bitcoin` will add `--bitcoin-node 127.0.0.1:18444` unless nodes are specified in dfx.json
+
 ### chore: Bump cdk to 0.19 in project template.
 
 ## Dependencies
 
 ### Replica
 
-Updated replica to elected commit 03783153d1596a81311b38cc1602063c400a3c8a.
-This incorporates the following executed proposals:
+Updated replica to elected commit 724ae4101bfdd8d4443126a6a8b1ec5ca9b68a12.
 
+This incorporates the following executed proposals:
+- [139570](https://dashboard.internetcomputer.org/proposal/139570)
+- [139480](https://dashboard.internetcomputer.org/proposal/139480)
+- [139403](https://dashboard.internetcomputer.org/proposal/139403)
 - [139317](https://dashboard.internetcomputer.org/proposal/139317)
 - [139192](https://dashboard.internetcomputer.org/proposal/139192)
 - [139079](https://dashboard.internetcomputer.org/proposal/139079)